Variables adjustment, colors, and secrets
Added secrets.py file as it is needed to be seen by libraries, but the data is not used. Updated README.md to reflect this.

Added color definitions and changed color callouts to plaintext for easier readability.

Changed alarm variable from being a 0/1 int variable to a boolean variable.

Added green LED verification of clearing weather alerts, updated README.md to reflect this.